expose connection attribute
