+++ catris/server_and_client.py
@@ -5,7 +5,11 @@ import collections
 import itertools
 import logging
 import time
-from typing import TYPE_CHECKING
+
+try:
+    from websockets.server import WebSocketServerProtocol
+except ImportError:
+    WebSocketServerProtocol = None  # type: ignore
 
 from catris.ansi import (
     CLEAR_FROM_CURSOR_TO_END_OF_SCREEN,
@@ -21,109 +25,9 @@ from catris.ansi import (
     MOVE_CURSOR,
     SHOW_CURSOR,
 )
+from catris.connections import RawTCPConnection, WebSocketConnection
 from catris.lobby import Lobby
 from catris.views import AskNameView, CheckTerminalSizeView, TextEntryView, View
-
-try:
-    from websockets.server import WebSocketServerProtocol
-    from websockets.exceptions import WebSocketException
-except ImportError:
-    WebSocketServerProtocol = None  # type: ignore
-    WebSocketException = None  # type: ignore
-
-
-class _RawTCPConnection:
-    def __init__(
-        self, reader: asyncio.StreamReader, writer: asyncio.StreamWriter
-    ) -> None:
-        self._reader = reader
-        self._writer = writer
-
-    def get_ip(self) -> str:
-        return self._writer.get_extra_info("peername")[0]
-
-    def get_send_queue_size(self) -> int:
-        # https://github.com/python/typeshed/issues/5779
-        return self._writer.transport.get_write_buffer_size()  # type: ignore
-
-    def put_to_send_queue(self, data: bytes) -> None:
-        self._writer.write(data)
-
-    async def receive_bytes(self) -> bytes:
-        return await self._reader.read(100)
-
-    async def flush(self) -> None:
-        await self._writer.drain()
-
-    def close(self) -> None:
-        self._writer.transport.close()
-
-    def is_closing(self) -> bool:
-        return self._writer.transport.is_closing()
-
-
-class _WebSocketConnection:
-    def __init__(self, ws: WebSocketServerProtocol) -> None:
-        self._ws = ws
-        self._send_queue = bytearray()
-        self._send_task: asyncio.Task[None] | None = None
-
-    def get_ip(self) -> str:
-        return self._ws.transport.get_extra_info("peername")[0]
-
-    def get_send_queue_size(self) -> int:
-        return len(self._send_queue)
-
-    def put_to_send_queue(self, data: bytes) -> None:
-        self._send_queue.extend(data)
-        if self._send_task is None or self._send_task.done():
-            self._send_task = asyncio.create_task(self._send_from_queue())
-
-    async def _send_from_queue(self) -> None:
-        while self._send_queue:
-            data_to_send = bytes(self._send_queue)
-            self._send_queue.clear()
-
-            try:
-                await self._ws.send(data_to_send)
-            except WebSocketException as e:
-                # Ideally we would know what client this connection belongs to.
-                # But for raw TCP connections, asyncio's internals log a message
-                # without any extra info anyway, so that wouldn't help much.
-                self.close()
-                logging.warning(f"sending to websocket failed: {e}")
-
-    async def receive_bytes(self) -> bytes:
-        try:
-            result = await self._ws.recv()
-        except WebSocketException as e:
-            raise OSError(str(e)) from e
-
-        if isinstance(result, str):
-            raise OSError("client sent text, expected bytes")
-        return result
-
-    async def flush(self):
-        if self._send_task is not None:
-            try:
-                await self._send_task
-            except WebSocketException as e:
-                raise OSError(str(e)) from e
-
-    # Docs say: "For legacy reasons, close() completes in at most
-    # 5 * close_timeout seconds for clients and 4 * close_timeout
-    # for servers."
-    #
-    # A small close_timeout is set when creating the connection, so
-    # this shouldn't create many simultaneously running tasks.
-    def close(self) -> None:
-        asyncio.create_task(self._ws.close())
-
-    def is_closing(self) -> bool:
-        # Docs say: "Be aware that both open and closed are False during the
-        # opening and closing sequences."
-        return not self._ws.open
-
 
 class Server:
     def __init__(self, use_lobbies: bool) -> None:
@@ -140,7 +44,7 @@ class Server:
     async def _handle_any_connection(
         self, client: Client
     ) -> None:
-        ip = client._connection.get_ip()
+        ip = client.connection.get_ip()
         self._connection_ips.append((time.monotonic(), ip))
         one_min_ago = time.monotonic() - 60
         while self._connection_ips and self._connection_ips[0][0] < one_min_ago:
@@ -155,12 +59,12 @@ class Server:
         await client.handle()
 
     async def handle_raw_tcp_connection(self, reader: asyncio.StreamReader, writer: asyncio.StreamWriter) -> None:
-        client = Client(self, _RawTCPConnection(reader, writer))
+        client = Client(self, RawTCPConnection(reader, writer))
         client.log("New raw TCP connection")
         await self._handle_any_connection(client)
 
     async def handle_websocket_connection(self, ws: WebSocketServerProtocol) -> None:
-        client = Client(self, _WebSocketConnection(ws))
+        client = Client(self, WebSocketConnection(ws))
         client.log("New websocket connection")
         await self._handle_any_connection(client)
 
@@ -170,9 +74,9 @@ _id_counter = itertools.count(1)
 
 class Client:
     def __init__(
-        self, server: Server, connection: _RawTCPConnection | _WebSocketConnection
+        self, server: Server, connection: RawTCPConnection | WebSocketConnection
     ) -> None:
-        self._connection = connection
+        self.connection = connection
         self._client_id = next(_id_counter)
         self.server = server
         self._current_receive_task: asyncio.Task[bytes] | None = None
@@ -187,9 +91,6 @@ class Client:
 
         self.rotate_counter_clockwise = False
         self.lobby_id_hidden = False
-
-        # Web UI doesn't have any way to resize. It's always big enough.
-        self.user_can_resize_terminal = not isinstance(connection, _WebSocketConnection)
 
     def get_lobby_id_for_display(self) -> bytes:
         assert self.lobby is not None
@@ -259,10 +160,10 @@ class Client:
         self._send_bytes(to_send)
 
     def _send_bytes(self, b: bytes) -> None:
-        if self._connection.is_closing():
+        if self.connection.is_closing():
             return
 
-        self._connection.put_to_send_queue(b)
+        self.connection.put_to_send_queue(b)
 
         # Prevent filling the server's memory if client sends but never receives.
         # Usually send buffer is empty (0 bytes) because operating system has buffering too.
@@ -270,9 +171,9 @@ class Client:
         #
         # On 80x24 terminal with no colors, we send max 80*24 = 1920 bytes at a time.
         # There's some extra space for colors and bigger terminals.
-        if self._connection.get_send_queue_size() > 4 * 1024:
+        if self.connection.get_send_queue_size() > 4 * 1024:
             self.log("More than 4K of data in send buffer, disconnecting")
-            self._connection.close()
+            self.connection.close()
             # Closing isn't enough to stop receiving immediately.
             # At least not with raw TCP connections
             if self._current_receive_task is not None:
@@ -283,11 +184,11 @@ class Client:
         # Should no longer be necessary, but just in case...
         await asyncio.sleep(0)
 
-        if self._connection.is_closing():
+        if self.connection.is_closing():
             return None
 
         assert self._current_receive_task is None
-        self._current_receive_task = asyncio.create_task(self._connection.receive_bytes())
+        self._current_receive_task = asyncio.create_task(self.connection.receive_bytes())
         try:
             result = await asyncio.wait_for(self._current_receive_task, timeout=10 * 60)
         except asyncio.TimeoutError:
@@ -381,7 +282,7 @@ class Client:
             self._send_bytes(b"\r" + CLEAR_FROM_CURSOR_TO_END_OF_SCREEN + SHOW_CURSOR)
 
             try:
-                await asyncio.wait_for(self._connection.flush(), timeout=3)
+                await asyncio.wait_for(self.connection.flush(), timeout=3)
             except (OSError, asyncio.TimeoutError):
                 pass
-            self._connection.close()
+            self.connection.close()
